requirements.txt
streamlit>=1.31.0
replicate>=0.22.0
openai>=1.12.0
arize-phoenix>=0.1.0
openinference-instrumentation-openai>=0.1.0
requests>=2.31.0
anthropic